# contracts/sandbox-worker-openapi.yaml
openapi: 3.1.0
info:
  title: Qgents Workspace 与 Sandbox Worker 内部接口
  version: 0.5.0
  description: 内部控制面接口。调用方不能提交宿主机路径、Git 远端、凭证或任意 Git 参数。
paths:
  /internal/v1/git-resolutions:
    post:
      operationId: resolveGitRef
      summary: Resolve a controlled Git ref to an immutable commit SHA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repositoryId, ref]
              properties:
                repositoryId: {type: string, format: uuid}
                ref: {type: string, minLength: 1, maxLength: 512}
      responses:
        '200':
          description: Resolved commit
          content:
            application/json:
              schema:
                type: object
                required: [commitSha]
                properties:
                  commitSha: {type: string}
  /internal/v1/health:
    get:
      operationId: health
      responses:
        '200':
          description: Worker 存活
          content:
            application/json:
              schema: {type: object, properties: {status: {type: string}}}
  /internal/v1/workspaces/{workspaceId}:
    parameters: [{$ref: '#/components/parameters/WorkspaceId'}]
    put:
      operationId: provisionWorkspace
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/WorkspaceProvisionRequest'}}}}
      responses: {'200': {description: 已幂等准备 Workspace, content: {application/json: {schema: {$ref: '#/components/schemas/Workspace'}}}}, '409': {description: 同一编号的规格冲突}}
    get:
      operationId: getWorkspace
      responses: {'200': {description: Workspace 与真实 HEAD, content: {application/json: {schema: {$ref: '#/components/schemas/Workspace'}}}}}
    delete:
      operationId: deleteWorkspace
      responses: {'204': {description: 已删除或原本不存在}, '409': {description: Workspace 仍被 Sandbox 使用}}
  /internal/v1/workspaces/{workspaceId}/repositories/{repositoryId}/git/status:
    parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/RepositoryId'}]
    get:
      operationId: getWorkspaceGitStatus
      responses: {'200': {description: 结构化状态, content: {application/json: {schema: {$ref: '#/components/schemas/GitStatus'}}}}}
  /internal/v1/workspaces/{workspaceId}/repositories/{repositoryId}/test-snapshots/{snapshotWorkspaceId}:
    parameters:
      - {$ref: '#/components/parameters/WorkspaceId'}
      - {$ref: '#/components/parameters/RepositoryId'}
      - {name: snapshotWorkspaceId, in: path, required: true, schema: {type: string, format: uuid}}
      - {name: projectId, in: query, required: true, schema: {type: string, format: uuid}}
    post:
      operationId: createWorkspaceTestSnapshot
      description: 将当前未提交工作树以受控 patch 固化到独立临时 Workspace；相同快照编号可幂等重试。
      responses:
        '200': {description: 已固化且哈希校验通过的隔离 Workspace, content: {application/json: {schema: {$ref: '#/components/schemas/Workspace'}}}}
        '409': {description: HEAD 或快照哈希不一致}
  /internal/v1/workspaces/{workspaceId}/repositories/{repositoryId}/git/diff:
    parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/RepositoryId'}]
    post:
      operationId: createWorkspaceGitDiff
      requestBody: {required: false, content: {application/json: {schema: {type: object, additionalProperties: false}}}}
      responses: {'200': {description: 包含未跟踪文件的完整 patch, content: {application/json: {schema: {$ref: '#/components/schemas/GitDiff'}}}}, '413': {description: Diff 超过上限}}
  /internal/v1/workspaces/{workspaceId}/repositories/{repositoryId}/git/commit:
    parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/RepositoryId'}]
    post:
      operationId: commitWorkspaceGitDiff
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/GitCommitRequest'}}}}
      responses: {'200': {description: 真实 Commit, content: {application/json: {schema: {$ref: '#/components/schemas/GitCommit'}}}}, '409': {description: Workspace 在用或审查快照不匹配}}
  /internal/v1/workspaces/{workspaceId}/repositories/{repositoryId}/git/push:
    parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/RepositoryId'}]
    post:
      operationId: pushWorkspaceBranch
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/GitPushRequest'}}}}
      responses: {'200': {description: 推送并核验成功, content: {application/json: {schema: {$ref: '#/components/schemas/GitPush'}}}}, '422': {description: origin 未配置或推送失败}}
  /internal/v1/git-stores/{repositoryId}/sync:
    parameters: [{$ref: '#/components/parameters/RepositoryId'}]
    post:
      operationId: syncGitStore
      description: 仅供主后端根据已绑定仓库元数据调用；不得由前端传入远程地址或凭据。
      parameters:
        - {name: Authorization, in: header, required: true, schema: {type: string}, description: 与 Worker 配置一致的内部服务 Bearer Token}
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/GitStoreSyncRequest'}}}}
      responses:
        '200': {description: bare Git Store 已初始化并完成指定分支同步, content: {application/json: {schema: {$ref: '#/components/schemas/GitStoreSync'}}}}
        '409': {description: 远程 HEAD 与预期 SHA 不一致}
        '422': {description: GitHub URL、分支或本地 Git Store 不合法}
  /internal/v1/sandboxes:
    post:
      operationId: createSandbox
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateSandboxRequest'}}}}
      responses: {'201': {description: Sandbox 已创建, content: {application/json: {schema: {$ref: '#/components/schemas/Sandbox'}}}}}
  /internal/v1/sandboxes/{sandboxId}:
    parameters: [{$ref: '#/components/parameters/SandboxId'}]
    get: {operationId: getSandbox, responses: {'200': {description: Sandbox 状态, content: {application/json: {schema: {$ref: '#/components/schemas/Sandbox'}}}}}}
    delete: {operationId: destroySandbox, responses: {'204': {description: Sandbox 已销毁，Workspace 保留}}}
  /internal/v1/sandboxes/{sandboxId}/lease/renew:
    parameters: [{$ref: '#/components/parameters/SandboxId'}]
    post:
      operationId: renewSandboxLease
      parameters:
        - {name: ttlSeconds, in: query, required: false, schema: {type: integer, format: int64, minimum: 1}}
      responses: {'200': {description: 租约已续期, content: {application/json: {schema: {$ref: '#/components/schemas/Sandbox'}}}}}
  /internal/v1/sandboxes/{sandboxId}/tool-executions:
    parameters: [{$ref: '#/components/parameters/SandboxId'}]
    post:
      operationId: executeTool
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ToolExecutionRequest'}}}}
      responses: {'202': {description: 已入队, content: {application/json: {schema: {$ref: '#/components/schemas/ToolExecution'}}}}}
  /internal/v1/tool-executions/{executionId}:
    parameters: [{$ref: '#/components/parameters/ExecutionId'}]
    get: {operationId: getToolExecution, responses: {'200': {description: 执行结果, content: {application/json: {schema: {$ref: '#/components/schemas/ToolExecution'}}}}}}
  /internal/v1/tool-executions/{executionId}/logs:
    parameters: [{$ref: '#/components/parameters/ExecutionId'}]
    get:
      operationId: getToolExecutionLogs
      parameters:
        - {name: after, in: query, schema: {type: integer, format: int64, minimum: 0, default: 0}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 1000, default: 200}}
      responses: {'200': {description: 有序增量日志, content: {application/json: {schema: {$ref: '#/components/schemas/ExecutionLogs'}}}}}
  /internal/v1/tool-executions/{executionId}/cancel:
    parameters: [{$ref: '#/components/parameters/ExecutionId'}]
    post: {operationId: cancelToolExecution, responses: {'200': {description: 已取消, content: {application/json: {schema: {$ref: '#/components/schemas/ToolExecution'}}}}}}
  /internal/v1/test-executions:
    post:
      operationId: executeTestsets
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/TestExecutionRequest'}}}}
      responses: {'200': {description: Controlled test results, content: {application/json: {schema: {$ref: '#/components/schemas/TestExecutionResponse'}}}}}
  /internal/v1/merge-previews:
    post:
      operationId: previewMerge
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/MergePreviewRequest'}}}}
      responses: {'200': {description: Read-only merge preview, content: {application/json: {schema: {$ref: '#/components/schemas/MergePreviewResponse'}}}}}
components:
  parameters:
    WorkspaceId: {name: workspaceId, in: path, required: true, schema: {type: string, format: uuid}}
    RepositoryId: {name: repositoryId, in: path, required: true, schema: {type: string, format: uuid}}
    SandboxId: {name: sandboxId, in: path, required: true, schema: {type: string, format: uuid}}
    ExecutionId: {name: executionId, in: path, required: true, schema: {type: string, format: uuid}}
  schemas:
    TestExecutionItemRequest:
      type: object
      required: [testsetId, command, timeoutSeconds, passRuleType, expectedExitCode]
      properties:
        testsetId: {type: string, format: uuid}
        command: {type: string, minLength: 1, maxLength: 4096}
        timeoutSeconds: {type: integer, minimum: 1, maximum: 3600}
        passRuleType: {type: string, enum: [EXIT_CODE]}
        expectedExitCode: {type: integer}
    TestExecutionRequest:
      type: object
      required: [executionId, projectId, repositoryId, testsets]
      properties:
        executionId: {type: string, format: uuid}
        projectId: {type: string, format: uuid}
        repositoryId: {type: string, format: uuid}
        workspaceId: {type: [string, 'null'], format: uuid}
        ref: {type: [string, 'null'], maxLength: 512}
        mergeSourceRef: {type: [string, 'null'], maxLength: 512, description: DryRun 时在一次性 checkout 中合并的受控源引用}
        testsets: {type: array, minItems: 1, maxItems: 32, items: {$ref: '#/components/schemas/TestExecutionItemRequest'}}
    TestExecutionResponse:
      type: object
      properties:
        executionId: {type: string, format: uuid}
        status: {type: string, enum: [PASSED, FAILED]}
        resolvedHeadCommit: {type: string}
        results: {type: array, items: {type: object, properties: {testsetId: {type: string, format: uuid}, status: {type: string}, exitCode: {type: [integer, 'null']}, durationMs: {type: integer}, failureCode: {type: [string, 'null']}}}}
    MergePreviewRequest:
      type: object
      required: [repositoryId, sourceRef, targetBranch]
      properties:
        repositoryId: {type: string, format: uuid}
        sourceRef: {type: string, minLength: 1, maxLength: 512}
        targetBranch: {type: string, minLength: 1, maxLength: 512}
    MergePreviewResponse:
      type: object
      properties:
        resolvedHeadCommit: {type: string}
        resolvedTargetCommit: {type: string}
        mergeable: {type: boolean}
        conflicts: {type: array, items: {type: string}}
    WorkspaceRepositoryRequest:
      type: object
      required: [repositoryId, baseRef, sourceBranch, workspacePath]
      properties:
        repositoryId: {type: string, format: uuid}
        baseRef: {type: string, minLength: 1, maxLength: 256, pattern: '^[A-Za-z0-9][A-Za-z0-9._/-]*$'}
        sourceBranch: {type: string, minLength: 1, maxLength: 256, pattern: '^[A-Za-z0-9][A-Za-z0-9._/-]*$'}
        workspacePath: {type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9][A-Za-z0-9._-]*$'}
    WorkspaceProvisionRequest:
      type: object
      required: [projectId, repositories]
      properties:
        projectId: {type: string, format: uuid}
        repositories: {type: array, minItems: 1, maxItems: 32, items: {$ref: '#/components/schemas/WorkspaceRepositoryRequest'}}
    WorkspaceRepository:
      allOf:
        - {$ref: '#/components/schemas/WorkspaceRepositoryRequest'}
        - {type: object, properties: {baseCommit: {type: string}, headCommit: {type: string}}}
    Workspace:
      type: object
      properties:
        id: {type: string, format: uuid}
        projectId: {type: string, format: uuid}
        storageKey: {type: string}
        status: {type: string, enum: [READY]}
        repositories: {type: array, items: {$ref: '#/components/schemas/WorkspaceRepository'}}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    GitChange:
      type: object
      properties: {indexStatus: {type: string}, worktreeStatus: {type: string}, path: {type: string}, originalPath: {type: [string, 'null']}}
    GitStatus:
      type: object
      properties: {branch: {type: string}, headCommit: {type: string}, clean: {type: boolean}, changes: {type: array, items: {$ref: '#/components/schemas/GitChange'}}}
    GitDiff:
      type: object
      properties: {headCommit: {type: string}, diffHash: {type: string}, patch: {type: string}}
    GitCommitRequest:
      type: object
      required: [expectedHeadCommit, expectedDiffHash, message, operationId]
      properties: {expectedHeadCommit: {type: string, minLength: 40, maxLength: 64, pattern: '^[0-9a-fA-F]{40,64}$'}, expectedDiffHash: {type: string, minLength: 71, maxLength: 71, pattern: '^sha256:[0-9a-f]{64}$'}, message: {type: string, minLength: 1, maxLength: 500}, operationId: {type: string, format: uuid, description: Commit 幂等操作ID}}
    GitCommit: {type: object, properties: {commitSha: {type: string}}}
    GitPushRequest:
      type: object
      required: [expectedHeadCommit, credentialGrantId]
      properties: {expectedHeadCommit: {type: string, minLength: 40, maxLength: 64, pattern: '^[0-9a-fA-F]{40,64}$'}, credentialGrantId: {type: string, minLength: 1, maxLength: 128}}
    GitPush: {type: object, properties: {branch: {type: string}, headCommit: {type: string}, verified: {type: boolean}}}
    GitStoreSyncRequest:
      type: object
      required: [repositoryUrl, remoteBranch, expectedHeadCommit, credentialGrantId]
      properties:
        repositoryUrl: {type: string, format: uri, description: 已绑定仓库推导出的 GitHub HTTPS URL；仅允许 github.com 且无 userinfo、query、fragment 或非 443 端口}
        remoteBranch: {type: string, minLength: 1, maxLength: 256, pattern: '^[A-Za-z0-9][A-Za-z0-9._/-]*$'}
        expectedHeadCommit: {type: string, minLength: 40, maxLength: 64, pattern: '^[0-9a-fA-F]{40,64}$'}
        credentialGrantId: {type: string, minLength: 1, maxLength: 128}
    GitStoreSync:
      type: object
      properties:
        repositoryId: {type: string, format: uuid}
        remoteBranch: {type: string}
        headCommit: {type: string}
        created: {type: boolean}
    ResourceLimits:
      type: object
      properties: {idleTtlSeconds: {type: integer, minimum: 1, maximum: 86400}, maxLifetimeSeconds: {type: integer, minimum: 1, maximum: 86400}, executionTimeoutSeconds: {type: integer, minimum: 1, maximum: 86400}}
    CreateSandboxRequest:
      type: object
      required: [sandboxId, taskRunId, workspaceStorageKey, imageProfile, repositoryIds]
      properties:
        sandboxId: {type: string, format: uuid}
        taskRunId: {type: string, format: uuid}
        workspaceStorageKey: {type: string, minLength: 47, maxLength: 47, pattern: '^workspaces/[0-9a-fA-F-]{36}$'}
        imageProfile: {type: string, minLength: 1, maxLength: 64, pattern: '^[a-z0-9][a-z0-9-]*$'}
        repositoryIds: {type: array, maxItems: 32, uniqueItems: true, items: {type: string, format: uuid}}
        limits: {$ref: '#/components/schemas/ResourceLimits'}
    Sandbox:
      type: object
      properties: {id: {type: string, format: uuid}, taskRunId: {type: string, format: uuid}, status: {type: string}, runtimeKind: {type: string}, createdAt: {type: string, format: date-time}, lastActiveAt: {type: string, format: date-time}, expiresAt: {type: string, format: date-time}, maxExpiresAt: {type: string, format: date-time}}
    ToolExecutionRequest:
      type: object
      required: [executionId, tool, arguments]
      properties:
        executionId: {type: string, format: uuid}
        repositoryId: {type: string, format: uuid}
        tool: {type: string, enum: [process.exec, file.read, file.list, file.search, file.write]}
        arguments: {type: object, maxProperties: 64}
        timeoutSeconds: {type: integer, minimum: 1, maximum: 3600}
    ToolExecution:
      type: object
      properties:
        id: {type: string, format: uuid}
        ownerWorkerId: {type: string}
        sandboxId: {type: string, format: uuid}
        repositoryId: {type: [string, 'null'], format: uuid}
        tool: {type: string}
        status: {type: string, enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, TIMED_OUT, CANCELLED, INTERRUPTED]}
        exitCode: {type: [integer, 'null']}
        result: {type: object}
        failureReason: {type: [string, 'null']}
        createdAt: {type: string, format: date-time}
        startedAt: {type: [string, 'null'], format: date-time}
        finishedAt: {type: [string, 'null'], format: date-time}
    ExecutionLogEntry:
      type: object
      properties:
        sequence: {type: integer, format: int64}
        stream: {type: string, enum: [SYSTEM, STDOUT, STDERR]}
        content: {type: string}
        timestamp: {type: string, format: date-time}
    ExecutionLogs:
      type: object
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/ExecutionLogEntry'}}
        nextCursor: {type: integer, format: int64}